+++ pkg/commands/commands.go
@@ -9,9 +9,12 @@ import (
 	aliasalerts "github.com/fastly/cli/pkg/commands/alias/alerts"
 	aliasbackend "github.com/fastly/cli/pkg/commands/alias/backend"
 	aliasdictionaryentry "github.com/fastly/cli/pkg/commands/alias/dictionaryentry"
+	aliasdictionary "github.com/fastly/cli/pkg/commands/alias/dictionary"
 	aliashealthcheck "github.com/fastly/cli/pkg/commands/alias/healthcheck"
 	aliasimageoptimizerdefaults "github.com/fastly/cli/pkg/commands/alias/imageoptimizerdefaults"
 	aliaspurge "github.com/fastly/cli/pkg/commands/alias/purge"
+	aliasratelimit "github.com/fastly/cli/pkg/commands/alias/ratelimit"
+	aliasresourcelink "github.com/fastly/cli/pkg/commands/alias/resourcelink"
 	aliasserviceversion "github.com/fastly/cli/pkg/commands/alias/serviceversion"
 	aliasvcl "github.com/fastly/cli/pkg/commands/alias/vcl"
 	aliasvclcondition "github.com/fastly/cli/pkg/commands/alias/vcl/condition"
@@ -26,6 +29,7 @@ import (
 	"github.com/fastly/cli/pkg/commands/dashboard"
 	dashboardItem "github.com/fastly/cli/pkg/commands/dashboard/item"
 	"github.com/fastly/cli/pkg/commands/dictionary"
+	"github.com/fastly/cli/pkg/commands/dictionaryentry"
 	"github.com/fastly/cli/pkg/commands/domain"
 	"github.com/fastly/cli/pkg/commands/install"
 	"github.com/fastly/cli/pkg/commands/ip"
@@ -93,8 +97,6 @@ import (
 	"github.com/fastly/cli/pkg/commands/pop"
 	"github.com/fastly/cli/pkg/commands/products"
 	"github.com/fastly/cli/pkg/commands/profile"
-	"github.com/fastly/cli/pkg/commands/ratelimit"
-	"github.com/fastly/cli/pkg/commands/resourcelink"
 	"github.com/fastly/cli/pkg/commands/secretstore"
 	"github.com/fastly/cli/pkg/commands/secretstoreentry"
 	"github.com/fastly/cli/pkg/commands/service"
@@ -103,10 +105,13 @@ import (
 	servicealert "github.com/fastly/cli/pkg/commands/service/alert"
 	servicebackend "github.com/fastly/cli/pkg/commands/service/backend"
 	servicedictionaryentry "github.com/fastly/cli/pkg/commands/service/dictionaryentry"
+	servicedictionary "github.com/fastly/cli/pkg/commands/service/dictionary"
 	servicedomain "github.com/fastly/cli/pkg/commands/service/domain"
 	servicehealthcheck "github.com/fastly/cli/pkg/commands/service/healthcheck"
 	serviceimageoptimizerdefaults "github.com/fastly/cli/pkg/commands/service/imageoptimizerdefaults"
 	servicepurge "github.com/fastly/cli/pkg/commands/service/purge"
+	serviceratelimit "github.com/fastly/cli/pkg/commands/service/ratelimit"
+	serviceresourcelink "github.com/fastly/cli/pkg/commands/service/resourcelink"
 	servicevcl "github.com/fastly/cli/pkg/commands/service/vcl"
 	servicevclcondition "github.com/fastly/cli/pkg/commands/service/vcl/condition"
 	servicevclcustom "github.com/fastly/cli/pkg/commands/service/vcl/custom"
@@ -203,6 +208,12 @@ func Define( // nolint:revive // function-length
 	dictionaryDescribe := dictionary.NewDescribeCommand(dictionaryCmdRoot.CmdClause, data)
 	dictionaryList := dictionary.NewListCommand(dictionaryCmdRoot.CmdClause, data)
 	dictionaryUpdate := dictionary.NewUpdateCommand(dictionaryCmdRoot.CmdClause, data)
+	dictionaryEntryCmdRoot := dictionaryentry.NewRootCommand(app, data)
+	dictionaryEntryCreate := dictionaryentry.NewCreateCommand(dictionaryEntryCmdRoot.CmdClause, data)
+	dictionaryEntryDelete := dictionaryentry.NewDeleteCommand(dictionaryEntryCmdRoot.CmdClause, data)
+	dictionaryEntryDescribe := dictionaryentry.NewDescribeCommand(dictionaryEntryCmdRoot.CmdClause, data)
+	dictionaryEntryList := dictionaryentry.NewListCommand(dictionaryEntryCmdRoot.CmdClause, data)
+	dictionaryEntryUpdate := dictionaryentry.NewUpdateCommand(dictionaryEntryCmdRoot.CmdClause, data)
 	domainCmdRoot := domain.NewRootCommand(app, data)
 	domainCreate := domain.NewCreateCommand(domainCmdRoot.CmdClause, data)
 	domainDelete := domain.NewDeleteCommand(domainCmdRoot.CmdClause, data)
@@ -559,18 +570,6 @@ func Define( // nolint:revive // function-length
 	profileSwitch := profile.NewSwitchCommand(profileCmdRoot.CmdClause, data, ssoCmdRoot)
 	profileToken := profile.NewTokenCommand(profileCmdRoot.CmdClause, data)
 	profileUpdate := profile.NewUpdateCommand(profileCmdRoot.CmdClause, data, ssoCmdRoot)
-	rateLimitCmdRoot := ratelimit.NewRootCommand(app, data)
-	rateLimitCreate := ratelimit.NewCreateCommand(rateLimitCmdRoot.CmdClause, data)
-	rateLimitDelete := ratelimit.NewDeleteCommand(rateLimitCmdRoot.CmdClause, data)
-	rateLimitDescribe := ratelimit.NewDescribeCommand(rateLimitCmdRoot.CmdClause, data)
-	rateLimitList := ratelimit.NewListCommand(rateLimitCmdRoot.CmdClause, data)
-	rateLimitUpdate := ratelimit.NewUpdateCommand(rateLimitCmdRoot.CmdClause, data)
-	resourcelinkCmdRoot := resourcelink.NewRootCommand(app, data)
-	resourcelinkCreate := resourcelink.NewCreateCommand(resourcelinkCmdRoot.CmdClause, data)
-	resourcelinkDelete := resourcelink.NewDeleteCommand(resourcelinkCmdRoot.CmdClause, data)
-	resourcelinkDescribe := resourcelink.NewDescribeCommand(resourcelinkCmdRoot.CmdClause, data)
-	resourcelinkList := resourcelink.NewListCommand(resourcelinkCmdRoot.CmdClause, data)
-	resourcelinkUpdate := resourcelink.NewUpdateCommand(resourcelinkCmdRoot.CmdClause, data)
 	secretstoreCmdRoot := secretstore.NewRootCommand(app, data)
 	secretstoreCreate := secretstore.NewCreateCommand(secretstoreCmdRoot.CmdClause, data)
 	secretstoreDescribe := secretstore.NewDescribeCommand(secretstoreCmdRoot.CmdClause, data)
@@ -614,6 +613,12 @@ func Define( // nolint:revive // function-length
 	serviceauthDescribe := serviceauth.NewDescribeCommand(serviceauthCmdRoot.CmdClause, data)
 	serviceauthList := serviceauth.NewListCommand(serviceauthCmdRoot.CmdClause, data)
 	serviceauthUpdate := serviceauth.NewUpdateCommand(serviceauthCmdRoot.CmdClause, data)
+	servicedictionaryCmdRoot := servicedictionary.NewRootCommand(serviceCmdRoot.CmdClause, data)
+	servicedictionaryCreate := servicedictionary.NewCreateCommand(servicedictionaryCmdRoot.CmdClause, data)
+	servicedictionaryDelete := servicedictionary.NewDeleteCommand(servicedictionaryCmdRoot.CmdClause, data)
+	servicedictionaryDescribe := servicedictionary.NewDescribeCommand(servicedictionaryCmdRoot.CmdClause, data)
+	servicedictionaryList := servicedictionary.NewListCommand(servicedictionaryCmdRoot.CmdClause, data)
+	servicedictionaryUpdate := servicedictionary.NewUpdateCommand(servicedictionaryCmdRoot.CmdClause, data)
 	servicevclCmdRoot := servicevcl.NewRootCommand(serviceCmdRoot.CmdClause, data)
 	servicevclDescribe := servicevcl.NewDescribeCommand(servicevclCmdRoot.CmdClause, data)
 	servicevclConditionCmdRoot := servicevclcondition.NewRootCommand(servicevclCmdRoot.CmdClause, data)
@@ -671,6 +676,18 @@ func Define( // nolint:revive // function-length
 	serviceimageoptimizerdefaultsCmdRoot := serviceimageoptimizerdefaults.NewRootCommand(serviceCmdRoot.CmdClause, data)
 	serviceimageoptimizerdefaultsGet := serviceimageoptimizerdefaults.NewGetCommand(serviceimageoptimizerdefaultsCmdRoot.CmdClause, data)
 	serviceimageoptimizerdefaultsUpdate := serviceimageoptimizerdefaults.NewUpdateCommand(serviceimageoptimizerdefaultsCmdRoot.CmdClause, data)
+	serviceratelimitCmdRoot := serviceratelimit.NewRootCommand(serviceCmdRoot.CmdClause, data)
+	serviceratelimitCreate := serviceratelimit.NewCreateCommand(serviceratelimitCmdRoot.CmdClause, data)
+	serviceratelimitDelete := serviceratelimit.NewDeleteCommand(serviceratelimitCmdRoot.CmdClause, data)
+	serviceratelimitDescribe := serviceratelimit.NewDescribeCommand(serviceratelimitCmdRoot.CmdClause, data)
+	serviceratelimitList := serviceratelimit.NewListCommand(serviceratelimitCmdRoot.CmdClause, data)
+	serviceratelimitUpdate := serviceratelimit.NewUpdateCommand(serviceratelimitCmdRoot.CmdClause, data)
+	serviceresourcelinkCmdRoot := serviceresourcelink.NewRootCommand(serviceCmdRoot.CmdClause, data)
+	serviceresourcelinkCreate := serviceresourcelink.NewCreateCommand(serviceresourcelinkCmdRoot.CmdClause, data)
+	serviceresourcelinkDelete := serviceresourcelink.NewDeleteCommand(serviceresourcelinkCmdRoot.CmdClause, data)
+	serviceresourcelinkDescribe := serviceresourcelink.NewDescribeCommand(serviceresourcelinkCmdRoot.CmdClause, data)
+	serviceresourcelinkList := serviceresourcelink.NewListCommand(serviceresourcelinkCmdRoot.CmdClause, data)
+	serviceresourcelinkUpdate := serviceresourcelink.NewUpdateCommand(serviceresourcelinkCmdRoot.CmdClause, data)
 	statsCmdRoot := stats.NewRootCommand(app, data)
 	statsHistorical := stats.NewHistoricalCommand(statsCmdRoot.CmdClause, data)
 	statsRealtime := stats.NewRealtimeCommand(statsCmdRoot.CmdClause, data)
@@ -738,6 +755,12 @@ func Define( // nolint:revive // function-length
 	aliasDictionaryEntryDescribe := aliasdictionaryentry.NewDescribeCommand(aliasDictionaryEntryRoot.CmdClause, data)
 	aliasDictionaryEntryList := aliasdictionaryentry.NewListCommand(aliasDictionaryEntryRoot.CmdClause, data)
 	aliasDictionaryEntryUpdate := aliasdictionaryentry.NewUpdateCommand(aliasDictionaryEntryRoot.CmdClause, data)
+	aliasDictionaryRoot := aliasdictionary.NewRootCommand(app, data)
+	aliasDictionaryCreate := aliasdictionary.NewCreateCommand(aliasDictionaryRoot.CmdClause, data)
+	aliasDictionaryDelete := aliasdictionary.NewDeleteCommand(aliasDictionaryRoot.CmdClause, data)
+	aliasDictionaryDescribe := aliasdictionary.NewDescribeCommand(aliasDictionaryRoot.CmdClause, data)
+	aliasDictionaryList := aliasdictionary.NewListCommand(aliasDictionaryRoot.CmdClause, data)
+	aliasDictionaryUpdate := aliasdictionary.NewUpdateCommand(aliasDictionaryRoot.CmdClause, data)
 	aliasHealthcheckRoot := aliashealthcheck.NewRootCommand(app, data)
 	aliasHealthcheckCreate := aliashealthcheck.NewCreateCommand(aliasHealthcheckRoot.CmdClause, data)
 	aliasHealthcheckDelete := aliashealthcheck.NewDeleteCommand(aliasHealthcheckRoot.CmdClause, data)
@@ -767,6 +790,18 @@ func Define( // nolint:revive // function-length
 	aliasACLEntryDescribe := aliasaclentry.NewDescribeCommand(aliasACLEntryRoot.CmdClause, data)
 	aliasACLEntryList := aliasaclentry.NewListCommand(aliasACLEntryRoot.CmdClause, data)
 	aliasACLEntryUpdate := aliasaclentry.NewUpdateCommand(aliasACLEntryRoot.CmdClause, data)
+	aliasRateLimitRoot := aliasratelimit.NewRootCommand(app, data)
+	aliasRateLimitCreate := aliasratelimit.NewCreateCommand(aliasRateLimitRoot.CmdClause, data)
+	aliasRateLimitDelete := aliasratelimit.NewDeleteCommand(aliasRateLimitRoot.CmdClause, data)
+	aliasRateLimitDescribe := aliasratelimit.NewDescribeCommand(aliasRateLimitRoot.CmdClause, data)
+	aliasRateLimitList := aliasratelimit.NewListCommand(aliasRateLimitRoot.CmdClause, data)
+	aliasRateLimitUpdate := aliasratelimit.NewUpdateCommand(aliasRateLimitRoot.CmdClause, data)
+	aliasResourceLinkRoot := aliasresourcelink.NewRootCommand(app, data)
+	aliasResourceLinkCreate := aliasresourcelink.NewCreateCommand(aliasResourceLinkRoot.CmdClause, data)
+	aliasResourceLinkDelete := aliasresourcelink.NewDeleteCommand(aliasResourceLinkRoot.CmdClause, data)
+	aliasResourceLinkDescribe := aliasresourcelink.NewDescribeCommand(aliasResourceLinkRoot.CmdClause, data)
+	aliasResourceLinkList := aliasresourcelink.NewListCommand(aliasResourceLinkRoot.CmdClause, data)
+	aliasResourceLinkUpdate := aliasresourcelink.NewUpdateCommand(aliasResourceLinkRoot.CmdClause, data)
 	aliasVclRoot := aliasvcl.NewRootCommand(app, data)
 	aliasVclDescribe := aliasvcl.NewDescribeCommand(aliasVclRoot.CmdClause, data)
 	aliasVclConditionRoot := aliasvclcondition.NewRootCommand(aliasVclRoot.CmdClause, data)
@@ -854,6 +889,12 @@ func Define( // nolint:revive // function-length
 		dictionaryDescribe,
 		dictionaryList,
 		dictionaryUpdate,
+		dictionaryEntryCmdRoot,
+		dictionaryEntryCreate,
+		dictionaryEntryDelete,
+		dictionaryEntryDescribe,
+		dictionaryEntryList,
+		dictionaryEntryUpdate,
 		domainCmdRoot,
 		domainCreate,
 		domainDelete,
@@ -1206,18 +1247,6 @@ func Define( // nolint:revive // function-length
 		profileSwitch,
 		profileToken,
 		profileUpdate,
-		rateLimitCmdRoot,
-		rateLimitCreate,
-		rateLimitDelete,
-		rateLimitDescribe,
-		rateLimitList,
-		rateLimitUpdate,
-		resourcelinkCmdRoot,
-		resourcelinkCreate,
-		resourcelinkDelete,
-		resourcelinkDescribe,
-		resourcelinkList,
-		resourcelinkUpdate,
 		secretstoreCreate,
 		secretstoreDescribe,
 		secretstoreDelete,
@@ -1258,6 +1287,12 @@ func Define( // nolint:revive // function-length
 		serviceauthDescribe,
 		serviceauthList,
 		serviceauthUpdate,
+		servicedictionaryCmdRoot,
+		servicedictionaryCreate,
+		servicedictionaryDelete,
+		servicedictionaryDescribe,
+		servicedictionaryList,
+		servicedictionaryUpdate,
 		servicevclCmdRoot,
 		servicevclDescribe,
 		servicevclConditionCmdRoot,
@@ -1306,6 +1341,18 @@ func Define( // nolint:revive // function-length
 		serviceimageoptimizerdefaultsCmdRoot,
 		serviceimageoptimizerdefaultsGet,
 		serviceimageoptimizerdefaultsUpdate,
+		serviceratelimitCmdRoot,
+		serviceratelimitCreate,
+		serviceratelimitDelete,
+		serviceratelimitDescribe,
+		serviceratelimitList,
+		serviceratelimitUpdate,
+		serviceresourcelinkCmdRoot,
+		serviceresourcelinkCreate,
+		serviceresourcelinkDelete,
+		serviceresourcelinkDescribe,
+		serviceresourcelinkList,
+		serviceresourcelinkUpdate,
 		serviceVersionActivate,
 		serviceVersionClone,
 		serviceVersionCmdRoot,
@@ -1379,6 +1426,11 @@ func Define( // nolint:revive // function-length
 		aliasDictionaryEntryDescribe,
 		aliasDictionaryEntryList,
 		aliasDictionaryEntryUpdate,
+		aliasDictionaryCreate,
+		aliasDictionaryDelete,
+		aliasDictionaryDescribe,
+		aliasDictionaryList,
+		aliasDictionaryUpdate,
 		aliasHealthcheckCreate,
 		aliasHealthcheckDelete,
 		aliasHealthcheckDescribe,
@@ -1404,6 +1456,16 @@ func Define( // nolint:revive // function-length
 		aliasACLEntryDescribe,
 		aliasACLEntryList,
 		aliasACLEntryUpdate,
+		aliasRateLimitCreate,
+		aliasRateLimitDelete,
+		aliasRateLimitDescribe,
+		aliasRateLimitList,
+		aliasRateLimitUpdate,
+		aliasResourceLinkCreate,
+		aliasResourceLinkDelete,
+		aliasResourceLinkDescribe,
+		aliasResourceLinkList,
+		aliasResourceLinkUpdate,
 		aliasVclDescribe,
 		aliasVclConditionCreate,
 		aliasVclConditionDelete,